Change rental_meter_ez type from Float to Integer

+++ app/models/rental_event.py
@@ -33,6 +33,6 @@ class Event(Base):
     fuel_capacity = Column(Float)
     fuel_type = Column(String)
     item_name = Column(String)
-    rental_meter_ez = Column(Float)
+    rental_meter_ez = Column(Integer)
     expected_return_date = Column(DateTime)
     item_type = Column(String)
